Separate S3 bucket name validation. Import in generateCoreTemplate.js.
