+++ lib/plugins/aws/package/lib/generateCoreTemplate.js
@@ -4,8 +4,15 @@ const BbPromise = require('bluebird');
 const path = require('path');
 const _ = require('lodash');
 
+const validateS3BucketName = require('../../lib/validateS3BucketName');
+
 module.exports = {
   generateCoreTemplate() {
+    _.assign(
+      this,
+      validateS3BucketName
+    );
+
     this.serverless.service.provider
       .compiledCloudFormationTemplate = this.serverless.utils.readFileSync(
       path.join(this.serverless.config.serverlessPath,
